wip: resolving conflicts in execution test suite

+++ test/libsolidity/semanticTests/array/copying/array_copy_different_packing.sol
@@ -7,7 +7,7 @@ contract c {
         returns (bytes10 a, bytes10 b, bytes10 c, bytes10 d, bytes10 e)
     {
         data1 = new bytes8[](9);
-        for (uint256 i = 0; i < data1.length; ++i) data1[i] = bytes8(uint64(i));
+        for (uint i = 0; i < data1.length; ++i) data1[i] = bytes8(uint64(i));
         data2 = data1;
         a = data2[1];
         b = data2[2];
