+++ test/libsolidity/semanticTests/array/dynamic_arrays_in_storage.sol
@@ -1,39 +1,39 @@
 contract c {
     struct Data {
-        uint256 x;
-        uint256 y;
+        uint x;
+        uint y;
     }
     Data[] data;
-    uint256[] ids;
+    uint[] ids;
 
-    function setIDStatic(uint256 id) public {
+    function setIDStatic(uint id) public {
         ids[2] = id;
     }
 
-    function setID(uint256 index, uint256 id) public {
+    function setID(uint index, uint id) public {
         ids[index] = id;
     }
 
-    function setData(uint256 index, uint256 x, uint256 y) public {
+    function setData(uint index, uint x, uint y) public {
         data[index].x = x;
         data[index].y = y;
     }
 
-    function getID(uint256 index) public returns (uint256) {
+    function getID(uint index) public returns (uint) {
         return ids[index];
     }
 
-    function getData(uint256 index) public returns (uint256 x, uint256 y) {
+    function getData(uint index) public returns (uint x, uint y) {
         x = data[index].x;
         y = data[index].y;
     }
 
-    function getLengths() public returns (uint256 l1, uint256 l2) {
+    function getLengths() public returns (uint l1, uint l2) {
         l1 = data.length;
         l2 = ids.length;
     }
 
-    function setLengths(uint256 l1, uint256 l2) public {
+    function setLengths(uint l1, uint l2) public {
         while (data.length < l1) data.push();
         while (ids.length < l2) ids.push();
     }
@@ -43,13 +43,13 @@ contract c {
 // compileViaYul: also
 // ----
 // getLengths() -> 0, 0
-// setLengths(uint256,uint256): 48, 49 ->
+// setLengths(uint,uint): 48, 49 ->
 // getLengths() -> 48, 49
-// setIDStatic(uint256): 11 ->
-// getID(uint256): 2 -> 11
-// setID(uint256,uint256): 7, 8 ->
-// getID(uint256): 7 -> 8
-// setData(uint256,uint256,uint256): 7, 8, 9 ->
-// setData(uint256,uint256,uint256): 8, 10, 11 ->
-// getData(uint256): 7 -> 8, 9
-// getData(uint256): 8 -> 10, 11
+// setIDStatic(uint): 11 ->
+// getID(uint): 2 -> 11
+// setID(uint,uint): 7, 8 ->
+// getID(uint): 7 -> 8
+// setData(uint,uint,uint): 7, 8, 9 ->
+// setData(uint,uint,uint): 8, 10, 11 ->
+// getData(uint): 7 -> 8, 9
+// getData(uint): 8 -> 10, 11
