+++ test/libsolidity/semanticTests/array/indexAccess/bytes_memory_index_access.sol
@@ -1,7 +1,7 @@
 contract Test {
-    function set(bytes memory _data, uint256 i)
+    function set(bytes memory _data, uint i)
         public
-        returns (uint256 l, bytes1 c)
+        returns (uint l, bytes1 c)
     {
         l = _data.length;
         c = _data[i];
@@ -10,4 +10,4 @@ contract Test {
 // ====
 // compileViaYul: also
 // ----
-// set(bytes,uint256): 0x40, 0x03, 0x08, "abcdefgh" -> 0x08, "d"
+// set(bytes,uint): 0x40, 0x03, 0x08, "abcdefgh" -> 0x08, "d"
