+++ test/libsolidity/semanticTests/integer/small_signed_types.sol
@@ -1,5 +1,5 @@
 contract test {
-    function run() public returns(int256 y) {
+    function run() public returns(int y) {
         return -int32(10) * -int64(20);
     }
 }
