+++ test/libsolidity/semanticTests/modifiers/function_modifier_multiple_times.sol
@@ -1,15 +1,15 @@
 contract C {
-    uint256 public a;
-    modifier mod(uint256 x) {
+    uint public a;
+    modifier mod(uint x) {
         a += x;
         _;
     }
 
-    function f(uint256 x) public mod(2) mod(5) mod(x) returns (uint256) {
+    function f(uint x) public mod(2) mod(5) mod(x) returns (uint) {
         return a;
     }
 }
 
 // ----
-// f(uint256): 3 -> 10
+// f(uint): 3 -> 10
 // a() -> 10
