+++ test/libsolidity/semanticTests/storage/complex_accessors.sol
@@ -1,8 +1,8 @@
 contract test {
-    mapping(uint256 => bytes4) public to_string_map;
-    mapping(uint256 => bool) public to_bool_map;
-    mapping(uint256 => uint256) public to_uint_map;
-    mapping(uint256 => mapping(uint256 => uint256)) public to_multiple_map;
+    mapping(uint => bytes4) public to_string_map;
+    mapping(uint => bool) public to_bool_map;
+    mapping(uint => uint) public to_uint_map;
+    mapping(uint => mapping(uint => uint)) public to_multiple_map;
     constructor() {
         to_string_map[42] = "24";
         to_bool_map[42] = false;
@@ -13,7 +13,7 @@ contract test {
 // ====
 // compileViaYul: also
 // ----
-// to_string_map(uint256): 42 -> "24"
-// to_bool_map(uint256): 42 -> false
-// to_uint_map(uint256): 42 -> 12
-// to_multiple_map(uint256,uint256): 42, 23 -> 31
+// to_string_map(uint): 42 -> "24"
+// to_bool_map(uint): 42 -> false
+// to_uint_map(uint): 42 -> 12
+// to_multiple_map(uint,uint): 42, 23 -> 31
