+++ test/libsolidity/semanticTests/storage/state_smoke_test.sol
@@ -1,11 +1,11 @@
 contract test {
-    uint256 value1;
-    uint256 value2;
-    function get(uint8 which) public returns (uint256 value) {
+    uint value1;
+    uint value2;
+    function get(uint8 which) public returns (uint value) {
         if (which == 0) return value1;
         else return value2;
     }
-    function set(uint8 which, uint256 value) public {
+    function set(uint8 which, uint value) public {
         if (which == 0) value1 = value;
         else value2 = value;
     }
@@ -16,9 +16,9 @@ contract test {
 // ----
 // get(uint8): 0x00 -> 0
 // get(uint8): 0x01 -> 0
-// set(uint8,uint256): 0x00, 0x1234 ->
-// set(uint8,uint256): 0x01, 0x8765 ->
+// set(uint8,uint): 0x00, 0x1234 ->
+// set(uint8,uint): 0x01, 0x8765 ->
 // get(uint8): 0x00 -> 0x1234
 // get(uint8): 0x01 -> 0x8765
-// set(uint8,uint256): 0x00, 0x03 ->
+// set(uint8,uint): 0x00, 0x03 ->
 // get(uint8): 0x00 -> 0x03
